# tests/contract/openapi_baseline.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Standardized error envelope returned by the API exception handlers.",
        "properties": {
          "detail": {
            "description": "Human-readable error message.",
            "type": "string"
          },
          "error_code": {
            "description": "Stable machine-readable error code.",
            "type": "string"
          },
          "errors": {
            "description": "Field-level validation errors (present on 422 validation failures).",
            "items": {
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "error_code",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "SpeciesCreate": {
        "description": "Schema for creating a new species.",
        "properties": {
          "care_level": {
            "default": "beginner",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "title": "Care Level",
            "type": "string"
          },
          "common_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Common Name",
            "type": "string"
          },
          "feeding_frequency": {
            "default": 2,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Feeding Frequency",
            "type": "integer"
          },
          "food_types": {
            "items": {
              "enum": [
                "flakes",
                "pellets",
                "frozen",
                "live",
                "vegetables",
                "algae"
              ],
              "type": "string"
            },
            "title": "Food Types",
            "type": "array"
          },
          "id": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "portion_hint": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portion Hint"
          },
          "scientific_name": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scientific Name"
          },
          "water_type": {
            "default": "freshwater",
            "enum": [
              "freshwater",
              "saltwater",
              "brackish"
            ],
            "title": "Water Type",
            "type": "string"
          }
        },
        "required": [
          "common_name",
          "id"
        ],
        "title": "SpeciesCreate",
        "type": "object"
      },
      "SpeciesListResponse": {
        "description": "Paginated response for species list.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SpeciesResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "pages": {
            "description": "Calculate total number of pages.",
            "readOnly": true,
            "title": "Pages",
            "type": "integer"
          },
          "per_page": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Per Page",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "per_page",
          "pages"
        ],
        "title": "SpeciesListResponse",
        "type": "object"
      },
      "SpeciesResponse": {
        "description": "Response schema for species data.",
        "properties": {
          "care_level": {
            "title": "Care Level",
            "type": "string"
          },
          "common_name": {
            "title": "Common Name",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "feeding_frequency": {
            "title": "Feeding Frequency",
            "type": "integer"
          },
          "food_types": {
            "items": {
              "type": "string"
            },
            "title": "Food Types",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "metadata_": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "portion_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portion Hint"
          },
          "scientific_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scientific Name"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "water_type": {
            "title": "Water Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "common_name",
          "scientific_name",
          "image_url",
          "food_types",
          "feeding_frequency",
          "portion_hint",
          "care_level",
          "water_type",
          "created_at",
          "updated_at"
        ],
        "title": "SpeciesResponse",
        "type": "object"
      },
      "SpeciesUpdate": {
        "description": "Schema for partial species update. All fields are optional.",
        "properties": {
          "care_level": {
            "anyOf": [
              {
                "enum": [
                  "beginner",
                  "intermediate",
                  "advanced"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Care Level"
          },
          "common_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Name"
          },
          "feeding_frequency": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feeding Frequency"
          },
          "food_types": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "flakes",
                    "pellets",
                    "frozen",
                    "live",
                    "vegetables",
                    "algae"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Food Types"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "portion_hint": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portion Hint"
          },
          "scientific_name": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scientific Name"
          },
          "water_type": {
            "anyOf": [
              {
                "enum": [
                  "freshwater",
                  "saltwater",
                  "brackish"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Type"
          }
        },
        "title": "SpeciesUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Backend API for FishFeed aquarium management system",
    "title": "FishFeed API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/species": {
      "post": {
        "description": "Create a new fish species entry.\n\nRequires admin privileges. Invalidates species list cache.",
        "operationId": "create_species_endpoint_api_v1_admin_species_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeciesCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesResponse"
                }
              }
            },
            "description": "Species created successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Admin privileges required"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Species with this ID already exists"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create a new species",
        "tags": [
          "Admin Species"
        ]
      }
    },
    "/api/v1/admin/species/{species_id}": {
      "delete": {
        "description": "Delete a fish species from the database.\n\nRequires admin privileges. Invalidates all species cache.",
        "operationId": "delete_species_endpoint_api_v1_admin_species__species_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "species_id",
            "required": true,
            "schema": {
              "title": "Species Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Species deleted successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Admin privileges required"
          },
          "404": {
            "description": "Species not found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete a species",
        "tags": [
          "Admin Species"
        ]
      },
      "put": {
        "description": "Update an existing fish species.\n\nRequires admin privileges. All fields are optional. Invalidates cache.",
        "operationId": "update_species_endpoint_api_v1_admin_species__species_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "species_id",
            "required": true,
            "schema": {
              "title": "Species Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeciesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesResponse"
                }
              }
            },
            "description": "Species updated successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Admin privileges required"
          },
          "404": {
            "description": "Species not found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update a species",
        "tags": [
          "Admin Species"
        ]
      }
    },
...
    "/api/v1/species": {
      "get": {
        "description": "Get a paginated list of fish species.\n\nSupports filtering by care_level and water_type.\nResults are ordered by common name and cached for 1 hour.",
        "operationId": "list_species_endpoint_api_v1_species_get",
        "parameters": [
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number",
              "maximum": 1000000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 100,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "description": "Filter by care level",
            "in": "query",
            "name": "care_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "beginner",
                    "intermediate",
                    "advanced"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by care level",
              "title": "Care Level"
            }
          },
          {
            "description": "Filter by water type",
            "in": "query",
            "name": "water_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "freshwater",
                    "saltwater",
                    "brackish"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by water type",
              "title": "Water Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesListResponse"
                }
              }
            },
            "description": "Paginated list of species"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "List all species",
        "tags": [
          "Species"
        ]
      }
    },
    "/api/v1/species/popular": {
      "get": {
        "description": "Get top 20 popular species for onboarding.\n\nReturns a curated list of common aquarium fish species\nsuitable for beginners and hobbyists. Cached for 1 hour.",
        "operationId": "get_popular_species_endpoint_api_v1_species_popular_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SpeciesResponse"
                  },
                  "title": "Response Get Popular Species Endpoint Api V1 Species Popular Get",
                  "type": "array"
                }
              }
            },
            "description": "List of popular species for onboarding"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "Get popular species",
        "tags": [
          "Species"
        ]
      }
    },
    "/api/v1/species/search": {
      "get": {
        "description": "Search species by name using full-text search.\n\nSearches both common_name and scientific_name fields.\nMinimum query length is 2 characters. Results cached for 30 minutes.",
        "operationId": "search_species_endpoint_api_v1_species_search_get",
        "parameters": [
          {
            "description": "Search query",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Search query",
              "maxLength": 100,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SpeciesResponse"
                  },
                  "title": "Response Search Species Endpoint Api V1 Species Search Get",
                  "type": "array"
                }
              }
            },
            "description": "List of matching species"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Query too short (minimum 2 characters)"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "Search species",
        "tags": [
          "Species"
        ]
      }
    },
    "/api/v1/species/{species_id}": {
      "get": {
        "description": "Get detailed information about a specific fish species. Cached for 24 hours.",
        "operationId": "get_species_endpoint_api_v1_species__species_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "species_id",
            "required": true,
            "schema": {
              "title": "Species Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesResponse"
                }
              }
            },
            "description": "Species details"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Species not found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "Get species by ID",
        "tags": [
          "Species"
        ]
      }
    },
...
  }
}